+++ src/controllers/productController 2.js
@@ -1,16 +1,15 @@
 const productService = require("../services/productService");
 
 const showMain = async (req, res) => {
-  // const { product_id } = req.body;
 
-  await productService.showMain(req, res);
+  await productService.showMain();
 
   res.status(201).json({ message: "show main product" });
 };
 const showSpecificProduct = async (req, res) => {
   await userService.signUp(req, res);
 
-  res.status(201).json({ message: "SHOW SPECIFIC PRODUCT COMPLETE " });
+  res.status(201).json({ message: "SHOW SPECIFIC PRODUCT COMPLETE ", });
 };
 
 module.exports = {
