+++ src/models/productDao.js
@@ -1,12 +1,14 @@
 const dataSource = require("./dataSource");
-const productController = require("../controllers/productController");
-
-const showMain = async (req, res) => {
+const showMain = async () => {
   await dataSource.query(`SELECT * FROM PRODUCT, PRODUCT_IMAGE`);
 };
 
 const showSpecificProduct = async (productId) => {
-  await dataSource.query(
+  /*await dataSource.query(
     `SELECT * FROM PRODUCT, PRODUCT_IMAGE WHERE PRODUCT.ID = [$productId] `
-  );
+  );*/
+};
+module.exports = {
+  showMain,
+  showSpecificProduct,
 };
