+++ src/services/productService.js
@@ -1,26 +1,16 @@
-const productDao = require("../models/productDao");
+const { productDao } = require("../models/productDao");
+//const { throwError } = require("../utils/throwError");
+const showMain = async () => {
+  // query thread from db
+  const products = await productDao.showMain();
 
-const showMain = (req, res) => {
-  /*
-    // 비밀번호 number, 소문자, 대문자, 특수문자, 5자 이상
-    const passwordRegex = /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*\W)(?!.*).$/
-
-    if (!passwordRegex.test(password)) {
-        const error = new Error ("NO_CHARACTERS")
-        error.status = 400
-        throw error
-    }
-*/
+  return products;
 };
-const showSpecificProduct = (req, res) => {
-  /*
-    // 비밀번호 number, 소문자, 대문자, 특수문자, 5자 이상
-    const passwordRegex = /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*\W)(?!.*).$/
 
-    if (!passwordRegex.test(password)) {
-        const error = new Error ("NO_CHARACTERS")
-        error.status = 400
-        throw error
-    }
-*/
+const showSpecificProduct = async (req, res) => {
+  let productId = req.productId;
+};
+module.exports = {
+  showMain,
+  showSpecificProduct,
 };
